--- src/tests/function/mod.rs
@@ -1,3 +1,0 @@
-use crate::build_suite;
-
-build_suite!(function, cast, left_right, upper_lower);
