+++ src/tests/functions/aggregate/mod.rs
@@ -1,0 +1,3 @@
+use crate::build_suite;
+
+build_suite!(aggregate; generic_aggregate, group_by);
