+++ src/tests/functions/arithmetic/mod.rs
@@ -1,0 +1,3 @@
+use crate::build_suite;
+
+build_suite!(arithmetic; generic_arithmatic, blend_arithmatic);
