+++ src/tests/functions/mod.rs
@@ -1,0 +1,6 @@
+pub mod aggregate;
+pub mod arithmetic;
+
+use crate::build_suite;
+
+build_suite!(functions; cast, left_right, upper_lower);
