+++ src/tests/generic/mod.rs
@@ -1,3 +1,14 @@
 use crate::build_suite;
 
-build_suite!(generic, basic, insert);
+build_suite!(generic;
+    basic,
+    create_table,
+    drop_table,
+    insert,
+    join,
+    join_blend,
+    nested_select,
+    ordering,
+    sql_types,
+    filter
+);
