+++ src/tests/tester.rs
@@ -53,14 +53,13 @@ pub trait Tester<T: 'static + Debug, U: Store<T> + StoreMut<T> + AlterTable> {
 
 #[macro_export]
 macro_rules! build_suite {
-    ($suite: ident, $($module: ident),+) => {
+    ($suite: ident; $($module: ident),+) => {
         $(pub mod $module;)+
         #[macro_export]
         macro_rules! $suite {
             ($test: meta, $storage: ident) => {
                 macro_rules! wrap {
                     ($name: ident) => {
-                        #[feature(concat_idents)]
                         #[$test]
                         // Once possible concat_idents!($ suite, $ name) would be nice (rustlang/rust#12249, rustlang/rust#29599)
                         async fn $name() {
