+++ src/tests/validate/mod.rs
@@ -1,3 +1,3 @@
 use crate::build_suite;
 
-build_suite!(validate, types, unique);
+build_suite!(validate; types, unique);
